move some e2e tests to SIG respectively
fix typo

include the servicecatalog package

add framework file

update golint failures
