+++ hack/.golint_failures
@@ -816,6 +816,7 @@ test/e2e/network
 test/e2e/node
 test/e2e/scalability
 test/e2e/scheduling
+test/e2e/servicecatalog
 test/e2e/storage
 test/e2e/upgrades
 test/e2e/upgrades/apps
