+++ test/e2e/BUILD
@@ -28,6 +28,7 @@ go_test(
         "//test/e2e/node:go_default_library",
         "//test/e2e/scalability:go_default_library",
         "//test/e2e/scheduling:go_default_library",
+        "//test/e2e/servicecatalog:go_default_library",
         "//test/e2e/storage:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
@@ -50,13 +51,11 @@ go_library(
         "events.go",
         "example_cluster_dns.go",
         "examples.go",
-        "generated_clientset.go",
         "gke_local_ssd.go",
         "gke_node_pools.go",
         "limit_range.go",
         "network_partition.go",
         "pod_gc.go",
-        "podpreset.go",
         "pods.go",
         "pre_stop.go",
         "resource_quota.go",
@@ -86,11 +85,8 @@ go_library(
         "//vendor/github.com/onsi/ginkgo/config:go_default_library",
         "//vendor/github.com/onsi/ginkgo/reporters:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/k8s.io/api/batch/v1:go_default_library",
-        "//vendor/k8s.io/api/batch/v2alpha1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/rbac/v1beta1:go_default_library",
-        "//vendor/k8s.io/api/settings/v1alpha1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/fields:go_default_library",
@@ -148,6 +144,7 @@ filegroup(
         "//test/e2e/perftype:all-srcs",
         "//test/e2e/scalability:all-srcs",
         "//test/e2e/scheduling:all-srcs",
+        "//test/e2e/servicecatalog:all-srcs",
         "//test/e2e/storage:all-srcs",
         "//test/e2e/testing-manifests:all-srcs",
         "//test/e2e/upgrades:all-srcs",
